+++ rpicomponents/src/pcf8574.hpp
@@ -13,14 +13,14 @@ namespace rpicomponents {
 		* @param pcf_pin_no: The pcf pin (between min_pin_offset_ and max_pin_offset_) for which the pin mode should be set
 		* @param pin_mode: The pin_mode to be set for the pin
 		*/
-		void SetPinMode(int pcf_pin_no, int pin_mode) const;
+		void SetPinMode(int8_t pcf_pin_no, int8_t pin_mode) const;
 	public:
 		/*
 		* Constructor
 		* @param address: The i2c address of the pcf8574 (must be greater than 63)
 		* @param pin_base: The pin base of the i2c component
 		*/
-		PCF8574(int address, int pin_base = 64);
+		PCF8574(int16_t address, int16_t pin_base = 64);
 
 		/*
 		* Method that writes output to a pcf pin
@@ -28,7 +28,7 @@ namespace rpicomponents {
 		* @param value: The value to be written (0-1)
 		* @param pcf_pin_no: The pcf pin to be written to
 		*/
-		void WriteToPcfPin(int pcf_pin_no, int value) const;
+		void WriteToPcfPin(int8_t pcf_pin_no, int16_t value) const;
 
 		/*
 		* Method that reads the input to a pcf pin
@@ -36,19 +36,19 @@ namespace rpicomponents {
 		* @param pcf_pin_no: The pcf pin to be read
 		* @returns: the read digital pin value (0 to 1)
 		*/
-		int ReadFromPcfPin(int pcf_pin_no) const;
+		int16_t ReadFromPcfPin(int8_t pcf_pin_no) const;
 
 		/*
 		* Method that returns the amount of pins that can be read from
 		* @returns: the amount of read pins (8)
 		*/
-		int AmountReadPins() const;
+		int8_t AmountReadPins() const;
 
 		/*
 		* Method that returns the amount of pins that can be written to
 		* @returns: the amount of write pins (8)
 		*/
-		int AmountWritePins() const;
+		int8_t AmountWritePins() const;
 	};
 }
